shipsticksenv/requirements.txt
playwright==1.50.0
pytest==8.3.3
pytest-playwright==0.8.0
pytest-xdist==3.6.1
pytest-html==4.1.1
python-dotenv==1.0.1